Camera orientation fails in Portrait mode
When bypass_rotation is enabled and auto-rotate is on,
changing camera orientation from landscape to portrait
mode fails.

Whenever auto-rotate is turned on, windows manager
triggers informs camera app to update display and
display gets updated. Hwc gets the updated frames
and since bypass_rotation is enabled, it performs
the rotation again on this.

Both bypass_rotation and Auto_rotation can't be
applied at the sametime and need to prioritize
one of these.

Whenever bypass_rotation is enable and auto_rotate is
on, prioritize auto_rotate.

Tracked-On: OAM-108836
Signed-off-by: Singh, Sapna1 <sapna1.singh@intel.com>